requirements/main.txt
#
# This file is autogenerated by pip-compile
# Make changes in main.in, then run this to update:
#
#    pip-compile main.in
#
alembic==0.8.4
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
arrow==0.7.0
Babel==2.2.0
bcrypt==2.0.0
billiard==3.3.0.22        # via celery
bleach==1.4.2             # via readme-renderer
boto3==1.2.3
botocore==1.3.24          # via boto3
celery==3.1.20
certifi==2015.11.20.1
cffi==1.5.0               # via bcrypt, cryptography
click==6.2
contextlib2==0.5.1        # via raven
cryptography==1.2.2       # via pyopenssl
docutils==0.12            # via botocore, readme-renderer
elasticsearch-dsl==0.0.9
elasticsearch==2.2.0
google-api-python-client==1.4.2
hiredis==0.2.0
html5lib==0.9999999
httplib2==0.9.2           # via google-api-python-client, oauth2client
idna==2.0                 # via cryptography
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0           # via boto3, botocore
kombu==3.0.33             # via celery
Mako==1.0.3
MarkupSafe==0.23
msgpack-python==0.4.7
oauth2client==1.5.2       # via google-api-python-client
packaging==16.1
paginate==0.5.2
passlib==1.6.5
PasteDeploy==1.5.2
psycopg2==2.6.1
pyasn1-modules==0.0.8     # via oauth2client
pyasn1==0.1.9             # via cryptography, oauth2client, pyasn1-modules, rsa
pycparser==2.14           # via cffi
Pygments==2.1
pyOpenSSL==0.15.1
pyparsing==2.1.0          # via packaging
pyramid-jinja2==2.6.2
pyramid-multiauth==0.7.0
pyramid-rpc==0.7.2
pyramid-services==0.4
pyramid-tm==0.12.1
pyramid==1.6.1
python-dateutil==2.4.2    # via arrow, botocore, elasticsearch-dsl
python-editor==0.5        # via alembic
pytz==2015.7              # via babel, celery
raven==5.10.2
readme-renderer==0.7.0
redis==2.10.5
repoze.lru==0.6           # via pyramid
requests==2.9.1
rfc3986==0.3.1
rsa==3.3                  # via oauth2client
simplejson==3.8.1         # via uritemplate
six==1.10.0               # via bcrypt, bleach, cryptography, elasticsearch-dsl, google-api-python-client, html5lib, oauth2client, packaging, pyopenssl, python-dateutil, readme-renderer, structlog
sqlalchemy-citext==1.3.post0
sqlalchemy==1.0.11
structlog==16.0.0
transaction==1.4.4
translationstring==1.3    # via pyramid
uritemplate==0.6          # via google-api-python-client
urllib3==1.14             # via elasticsearch
venusian==1.0             # via pyramid, pyramid-rpc
WebOb==1.5.1
WTForms==2.1
zope.deprecation==4.1.2   # via pyramid, pyramid-jinja2
zope.interface==4.1.3     # via pyramid, pyramid-services, transaction, zope.sqlalchemy
zope.sqlalchemy==0.7.6

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools
